Refactor web review card components
