+++ apps/web/src/screens/review/useReviewScreenController.ts
@@ -46,13 +46,13 @@ import { useTransientMessage } from "../../useTransientMessage";
 import type { Card, FeedbackPromptEventType, FeedbackSubmissionRequest } from "../../types";
 import { handleRefreshLocalDataError } from "../shared/refreshLocalDataError";
 import { isCardFormStateDirty } from "../cards/form/CardForm";
-import type { ReviewEditorModalProps } from "./components/ReviewEditorModal";
+import type { ReviewEditorModalProps } from "./components/card/ReviewEditorModal";
 import type { ReviewPaneProps } from "./components/ReviewPane";
 import type { ReviewQueuePanelProps } from "./components/ReviewQueuePanel";
 import type { ReviewScreenHeaderProps } from "./components/ReviewScreenHeader";
 import { buildReviewButtonOptions, type ReviewButtonOption } from "./components/reviewRatingOptions";
 import { type LastSubmittedReview, type ReviewSubmitState } from "./components/reviewScreenTypes";
-import { useReviewCardEditor } from "./components/useReviewCardEditor";
+import { useReviewCardEditor } from "./components/card/useReviewCardEditor";
 import { useReviewScreenData, type ReviewSubmissionOutcome } from "./data/useReviewScreenData";
 import { useReviewFilterMenu } from "./filters/useReviewFilterMenu";
 import type { ReviewHardReminderDialogProps } from "./hardReminder/ReviewHardReminderDialog";
